+++ html/index.html
@@ -23,7 +23,7 @@
       <main>
         <div class="auto-rolling-news-bar">
           <div class="auto-rolling-news">
-            <ul id="left-rolling">
+            <ul data-type="left-rolling">
               <li class="current">
                 <span class="press">연합뉴스1</span>
                 <span class="news-title">[1111보] 김기현,안철수~~</span>
@@ -47,7 +47,7 @@
             </ul>
           </div>
           <div class="auto-rolling-news">
-            <ul id="right-rolling">
+            <ul data-type="right-rolling">
               <li class="current">
                 <span class="press">연합뉴스</span>
                 <span class="news-title">[속보1] 최고위원 본경선,~~</span>
