+++ js/main/changeView/changeView.js
@@ -1,5 +1,5 @@
 //뷰 타입 변경
-import { main_list_page, MIN_PAGE, MAX_PAGE } from "./gridView/gridView.js";
+import { main_list_page, MIN_PAGE, MAX_PAGE } from "../gridView/gridView.js";
 
 function getById(id) {
   return document.getElementById(id);
